+++ src/crypto/generichash/mod.rs
@@ -106,6 +106,7 @@ impl State {
 #[cfg(test)]
 mod test {
     use super::*;
+    use hex;
     #[cfg(not(feature = "std"))]
     use prelude::*;
 
@@ -162,7 +163,7 @@ mod test {
 
                 match line.len() {
                     0 => break,
-                    1...3 => continue,
+                    1..=3 => continue,
                     _ => {}
                 }
 
@@ -190,5 +191,26 @@ mod test {
             let result_hash = hasher.finalize().unwrap();
             assert!(result_hash.as_ref() == expected_hash.as_slice());
         }
+    }
+
+    #[test]
+    fn test_digest_equality() {
+        let data1 = [1, 2];
+        let data2 = [3, 4];
+
+        let h1 = {
+            let mut hasher = State::new(32, None).unwrap();
+            hasher.update(&data1).unwrap();
+            hasher.finalize().unwrap()
+        };
+
+        let h2 = {
+            let mut hasher = State::new(32, None).unwrap();
+            hasher.update(&data2).unwrap();
+            hasher.finalize().unwrap()
+        };
+
+        assert_eq!(h1, h1);
+        assert_ne!(h1, h2);
     }
 }