+++ src/crypto/hash/sha512.rs
@@ -24,6 +24,7 @@ hash_module!(
 #[cfg(test)]
 mod test {
     use super::*;
+    use hex;
     #[cfg(not(feature = "std"))]
     use prelude::*;